dont know whats going on

+++ Project_3/PenningTrap.hpp
@@ -15,11 +15,11 @@ public:
     double d_;            // characteristic dimension
     double ke_;           // coloumbs constant
     int n_;               // number of particles
-    int N_;               // number of time steps
+    double N_;               // number of time steps
     std::vector<Particle> particles_;       // to contain all the Particle objects in the Penning trap
 
     // constructor
-    PenningTrap(double B0, double V0, double d, double ke, int n, int N, arma::mat pos, arma::mat vel, arma::vec q_vec, arma::vec m_vec);
+    PenningTrap(double B0, double V0, double d, double ke, int n, double N, arma::mat pos, arma::mat vel, arma::vec q_vec, arma::vec m_vec);
 
     // class methods to be made
     arma::vec external_E_field(int i);      // external electric field at point r=(x,y,z)
